+++ src/components/text-note/footer/Footer.tsx
@@ -3,9 +3,9 @@ import { useMutation } from '@tanstack/react-query'
 import { useRouteContext } from '@tanstack/react-router'
 import { Heart } from 'lucide-react'
 import { useMemo, useState } from 'react'
+import useNostr from '@/lib/nostr/hooks/use-nostr'
+import { useNostrEvents } from '@/lib/nostr/hooks/use-nostr-events'
 import { ReactionEventSchema } from '@/lib/nostr/kinds/7'
-import useNostr from '@/lib/nostr/use-nostr'
-import { useNostrEvents } from '@/lib/nostr/use-nostr-events'
 import queryKeys from '@/lib/query-keys'
 import TextNoteForm from '../form/TextNoteForm'
 import Emoji from './Emoji'
