+++ src/components/text-note/form/TextNoteForm.tsx
@@ -1,9 +1,9 @@
 import { useMutation } from '@tanstack/react-query'
 import { SendHorizonal } from 'lucide-react'
 import { useState } from 'react'
+import useNostr from '@/lib/nostr/hooks/use-nostr'
 import { TextNoteEventSchema } from '@/lib/nostr/kinds/1'
 import { createEvent } from '@/lib/nostr/nip19'
-import useNostr from '@/lib/nostr/use-nostr'
 import { Button } from '@/shadcn-ui/components/ui/button'
 import { Textarea } from '@/shadcn-ui/components/ui/textarea'
 import ContentWarningForm from './ContentWarningForm'
