+++ src/routes/(app)/settings/relays.tsx
@@ -3,8 +3,8 @@ import { CloudDownload, CloudUpload } from 'lucide-react'
 import { useEffect, useRef, useState } from 'react'
 import { RelayForm } from '@/components/page/settings/relay/RelayForm'
 import { RelayTable } from '@/components/page/settings/relay/RelayTable'
+import useNostr from '@/lib/nostr/hooks/use-nostr'
 import { RelayListEventSchema, relayListQuery } from '@/lib/nostr/kinds/10002'
-import useNostr from '@/lib/nostr/use-nostr'
 import { readStore, writeStore } from '@/lib/store'
 import { Button } from '@/shadcn-ui/components/ui/button'
